user logs in from new device(private key and public key storage) case addressed
checks if get-api returns keys but they are not present in session storage. this is the case where user logs in from another device. so in this case we prompt the user for master password, fetches private key from get-key , decrypts it using master password and stores it in session storage. public key is stored from get-key as well.